Files/requirements.txt
aiohttp
aiofiles
requests
ipaddress
PyYAML